Link with standard C library by default.
Note: We do not have library support yet, so we link
with stubs.z80.pp.obj.

+++ src/pathname.c
@@ -47,7 +47,7 @@ char *pathname_get_dirname(const char *pathname)
 
 	sep = strrchr(pathname, '/');
 	if (sep != NULL) {
-		off = sep - pathname;
+		off = (size_t)(sep - pathname);
 		dirname = malloc(off + 1);
 		if (dirname == NULL)
 			return NULL;
